+++ src/main/java/org/generation/BlogPessoal/Controller/TemaController.java
@@ -1,6 +1,9 @@
 package org.generation.BlogPessoal.Controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.generation.BlogPessoal.model.Tema;
 import org.generation.BlogPessoal.repository.TemaRepository;
@@ -15,47 +18,52 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 @RestController
-@CrossOrigin(origins = "*", allowedHeaders = "*" )
+@CrossOrigin(origins = "*", allowedHeaders = "*")
 @RequestMapping("/tema")
 public class TemaController {
-   
-	@Autowired 
+
+	@Autowired
 	private TemaRepository repository;
-	
+
 	@GetMapping
-	public ResponseEntity<List<Tema>> getAll(){
+	public ResponseEntity<List<Tema>> getAll() {
 		return ResponseEntity.ok(repository.findAll());
 	}
-	
-	@GetMapping ("/{id}")
-	public ResponseEntity<Tema> getById(@PathVariable long id){
-		return repository.findById(id).map(resp -> ResponseEntity.ok(resp))
-				.orElse(ResponseEntity.notFound().build());
-		
-    }
-	
-	@GetMapping("/nome/{nome}")
-	public ResponseEntity<List <Tema>> getByName (@PathVariable String nome){
-		return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(nome));
-}
 
-    @PostMapping
-    public ResponseEntity<Tema> post (@RequestBody Tema tema){
-    	return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema));
-    }
-    
-    @PutMapping
-    public ResponseEntity<Tema> put(@RequestBody Tema tema){
-    	return ResponseEntity.ok(repository.save(tema));
-    }
-    
-    @DeleteMapping("/{id}")
-    public void delete(@PathVariable long id) {
-    	repository.deleteById(id);
-   
-    }
-    
- }
+	@GetMapping("/{id}")
+	public ResponseEntity<Tema> getById(@PathVariable long id) {
+		return repository.findById(id).map(resp -> ResponseEntity.ok(resp)).orElse(ResponseEntity.notFound().build());
+
+	}
+
+	@GetMapping("/nome/{nome}")
+	public ResponseEntity<List<Tema>> getByName(@PathVariable String nome) {
+		return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(nome));
+	}
+
+	@PostMapping
+	public ResponseEntity<Tema> post(@Valid @RequestBody Tema tema) {
+		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema));
+	}
+
+	@PutMapping
+	public ResponseEntity<Tema> put(@Valid @RequestBody Tema tema) {
+		return repository.findById(tema.getId())
+				.map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
+	}
+
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("/{id}")
+	public void delete(@PathVariable long id) {
+		Optional<Tema> tema = repository.findById(id);
+		if (tema.isEmpty())
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+		repository.deleteById(id);
+	}
+}
